[addons][pvr] remove deprecated kodi_vfs_types.h
The two structures currently used in it are moved to libXBMC_addon.h
and kodi_vfs_types.h deleted.
