+++ xbmc/addons/interfaces/AddonInterfaces.h
@@ -9,9 +9,9 @@
 
 #pragma once
 
-#include "addons/kodi-addon-dev-kit/include/kodi/libXBMC_addon.h"
-
 #include <stdint.h>
+
+struct AddonCB;
 
 namespace KODI
 {
